frontend.views.upload: move UploadStatus to library
